+++ CHANGES.md
@@ -14,6 +14,7 @@
 - Fixed broken instrumentation when custom sources directory is set [#1004](../../issues/1004)
 - Fixed `java.nio.file.FileAlreadyExistsException: instrumentCode/classpath.index` [#998](../../issues/998)
 - Fixed `Execution optimizations have been disabled for task ':jar' to ensure correctness` [#1000](../../issues/1000)
+- Fixed JaCoCo `Can't add different class with same name` exception when using code instrumentation [#1020](../../issues/1020)
 
 ## 1.6.0
 ### Added